# openapi.yaml
openapi: 3.0.3
info:
  title: MVP HLS API
  description: REST API for HLS video streaming platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
tags:
  - name: Videos
    description: Video upload and HLS streaming
paths:
  /api/videos/upload/signature:
    post:
      tags:
        - Videos
      summary: Create signed direct upload parameters
      description: >-
        Creates a video record and returns signed Cloudinary parameters so the frontend can upload the raw video
        directly to Cloudinary.
      operationId: createSignedVideoUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSignedVideoUploadRequest'
      responses:
        '201':
          description: Signed upload parameters created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedVideoUploadResponse'
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/videos:
    get:
      tags:
        - Videos
      summary: List videos
      operationId: listVideos
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/VideoStatus'
      responses:
        '200':
          description: Paginated list of videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/videos/{id}:
    get:
      tags:
        - Videos
      summary: Get a video by ID
      operationId: getVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      responses:
        '200':
          description: Video found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Videos
      summary: Delete a video
      operationId: deleteVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      responses:
        '204':
          description: Video deleted
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/videos/{id}/upload-complete:
    post:
      tags:
        - Videos
      summary: Complete direct video upload
      description: >-
        Verifies the Cloudinary upload response, stores the raw asset metadata, and queues the video for HLS
        transcoding.
      operationId: completeSignedVideoUpload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteSignedVideoUploadRequest'
      responses:
        '202':
          description: Upload completion accepted — processing queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoUploadResponse'
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Video is not in a state that can complete upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/videos/{id}/retry:
    post:
      tags:
        - Videos
      summary: Retry failed video processing
      operationId: retryVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      responses:
        '202':
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoUploadResponse'
        '400':
          description: Bad Request / Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Meta:
      type: object
      required:
        - total
        - page
        - limit
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
    ErrorResponse:
      type: object
      required:
        - statusCode
        - error
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          oneOf:
            - type: string
              example: Validation failed
            - type: array
              items:
                type: string
    VideoStatus:
      type: string
      enum:
        - uploaded
        - queued
        - processing
        - ready
        - failed
        - deleted
    VideoVariant:
      type: object
      required:
        - id
        - qualityLabel
        - width
        - height
        - bitrateKbps
      properties:
        id:
          type: string
          format: uuid
        qualityLabel:
          type: string
          example: 1080p
        width:
          type: integer
          example: 1920
        height:
          type: integer
          example: 1080
        bitrateKbps:
          type: integer
          example: 5000
        playlistUrl:
          type: string
          format: uri
          nullable: true
    Video:
      type: object
      required:
        - id
        - title
        - status
        - progress
        - variants
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
          example: My Video
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/VideoStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 75
        playbackUrl:
          type: string
          format: uri
          nullable: true
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        durationSec:
          type: number
          nullable: true
          example: 120.5
        width:
          type: integer
          nullable: true
          example: 1920
        height:
          type: integer
          nullable: true
          example: 1080
        errorMessage:
          type: string
          nullable: true
        variants:
          type: array
          items:
            $ref: '#/components/schemas/VideoVariant'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VideoUploadResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - id
            - status
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/VideoStatus'
    CreateSignedVideoUploadRequest:
      type: object
      required:
        - title
        - originalFilename
        - mimeType
        - sizeBytes
      properties:
        title:
          type: string
          maxLength: 255
          example: My Video
        description:
          type: string
          example: Optional description
        originalFilename:
          type: string
          maxLength: 255
          example: intro.mp4
        mimeType:
          type: string
          enum:
            - video/mp4
            - video/quicktime
            - video/x-msvideo
            - video/webm
            - video/x-matroska
          example: video/mp4
        sizeBytes:
          type: integer
          minimum: 1
          maximum: 524288000
          example: 10485760
    SignedVideoUploadResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - videoId
            - uploadUrl
            - publicId
            - apiKey
            - timestamp
            - signature
            - resourceType
            - maxFileSize
            - uploadParams
          properties:
            videoId:
              type: string
              format: uuid
            uploadUrl:
              type: string
              format: uri
              example: https://api.cloudinary.com/v1_1/demo/video/upload
            publicId:
              type: string
              example: mvp-hls/raw/video-id
            apiKey:
              type: string
              example: '1234567890'
            timestamp:
              type: integer
              example: 1781111111
            signature:
              type: string
              example: cloudinary-signature
            resourceType:
              type: string
              enum:
                - video
              example: video
            maxFileSize:
              type: integer
              maximum: 524288000
              example: 524288000
            uploadParams:
              type: object
              required:
                - public_id
                - timestamp
                - api_key
                - signature
                - overwrite
              properties:
                public_id:
                  type: string
                  example: mvp-hls/raw/video-id
                timestamp:
                  type: integer
                  example: 1781111111
                api_key:
                  type: string
                  example: '1234567890'
                signature:
                  type: string
                  example: cloudinary-signature
                overwrite:
                  type: boolean
                  enum:
                    - false
                  example: false
    CompleteSignedVideoUploadRequest:
      type: object
      required:
        - publicId
        - version
        - signature
        - secureUrl
      properties:
        publicId:
          type: string
          example: mvp-hls/raw/video-id
        version:
          type: integer
          minimum: 1
          example: 1781111111
        signature:
          type: string
          example: cloudinary-response-signature
        secureUrl:
          type: string
          format: uri
          example: https://res.cloudinary.com/demo/video/upload/v1781111111/mvp-hls/raw/video-id.mp4
    VideoResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Video'
    VideoListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Video'
        meta:
          $ref: '#/components/schemas/Meta'